优化常量
